Add player death Add Enemy attack Add colliders for map Rework map Add Shake for camera Add ammo stats Fixed enemy Spawn

+++ Assets/InternalAssets/Scripts/Interfaces/IDamaged.cs
@@ -1,4 +1,4 @@
 ﻿public interface IDamaged
 {
-    public void GetDamage(int damage);
+    public void GetDamage(float damage);
 }
